Feat category into items pages : modify server actions

+++ src/app/actions/drawings/admin.ts
@@ -71,7 +71,7 @@ export async function updateDrawing(
       const newFile = rawFormData.file as File;
       const title = rawFormData.title as string;
       if (newFile.size !== 0) {
-        deleteFile(dir, oldDraw.images[0].filename);
+        deleteFile(dir, oldDraw.imageFilename);
         fileInfo = await resizeAndSaveImage(newFile, title, dir);
       }
 
@@ -82,7 +82,7 @@ export async function updateDrawing(
                 id: Number(rawFormData.categoryId),
               },
             }
-          : oldDraw.categoryId !== null
+          : oldDraw.categoryId
             ? {
                 disconnect: {
                   id: oldDraw.categoryId,
@@ -122,7 +122,7 @@ export async function deleteDrawing(id: number) {
       where: { id },
     });
     if (drawing) {
-      deleteFile(dir, drawing.images[0].filename);
+      deleteFile(dir, drawing.imageFilename);
       await prisma.drawing.delete({
         where: {
           id,
@@ -138,9 +138,22 @@ export async function deleteDrawing(id: number) {
 
 export async function deleteCategoryDrawing(id: number) {
   try {
-    await prisma.drawingCategory.delete({
+    const cat = await prisma.drawingCategory.findUnique({
       where: { id },
     });
+
+    if (cat) {
+      const contentId = cat.categoryContentId;
+      if (contentId) {
+        await prisma.categoryContent.delete({
+          where: { id: contentId },
+        });
+      }
+
+      await prisma.drawingCategory.delete({
+        where: { id },
+      });
+    }
     revalidatePath("/admin/dessins");
     return { message: "Catégorie supprimée", isError: false };
   } catch (e) {
@@ -152,16 +165,26 @@ export async function createCategoryDrawing(
   prevState: { message: string; isError: boolean } | null,
   formData: FormData,
 ) {
-  try {
-    const value = formData.get("text") as string;
-    const key = transformValueToKey(value);
+  const rawFormData = Object.fromEntries(formData);
+  const value = rawFormData.value as string;
 
+  try {
     await prisma.drawingCategory.create({
       data: {
-        key,
+        key: transformValueToKey(value),
         value,
+        content: {
+          create: {
+            title: rawFormData.title as string,
+            text: rawFormData.text as string,
+            imageFilename: rawFormData.filename as string,
+            imageWidth: Number(rawFormData.width),
+            imageHeight: Number(rawFormData.height),
+          },
+        },
       },
     });
+
     revalidatePath("/admin/dessins");
     return { message: "Catégorie ajoutée", isError: false };
   } catch (e) {
@@ -173,19 +196,46 @@ export async function updateCategoryDrawing(
   prevState: { message: string; isError: boolean } | null,
   formData: FormData,
 ) {
-  try {
-    const rawFormData = Object.fromEntries(formData);
-    const id = Number(rawFormData.id);
-    const value = rawFormData.text as string;
-    const key = transformValueToKey(value);
+  const rawFormData = Object.fromEntries(formData);
+  const id = Number(rawFormData.id);
+  const value = rawFormData.value as string;
 
-    await prisma.drawingCategory.update({
+  try {
+    const oldCat = await prisma.drawingCategory.findUnique({
       where: { id },
-      data: {
-        key,
-        value,
-      },
+      include: { content: true },
     });
+
+    if (oldCat) {
+      let content;
+      const data = {
+        title: rawFormData.title as string,
+        text: rawFormData.text as string,
+        imageFilename: rawFormData.filename as string,
+        imageWidth: Number(rawFormData.width),
+        imageHeight: Number(rawFormData.height),
+      };
+
+      if (!oldCat.content) {
+        content = {
+          create: data,
+        };
+      } else {
+        content = {
+          update: data,
+        };
+      }
+
+      await prisma.drawingCategory.update({
+        where: { id },
+        data: {
+          key: transformValueToKey(value),
+          value,
+          content,
+        },
+      });
+    }
+
     revalidatePath("/admin/dessins");
     return { message: "Catégorie modifiée", isError: false };
   } catch (e) {
