+++ src/app/actions/paintings/admin.ts
@@ -71,7 +71,7 @@ export async function updatePainting(
       const newFile = rawFormData.file as File;
       const title = rawFormData.title as string;
       if (newFile.size > 0) {
-        deleteFile(dir, oldPaint.images[0].filename);
+        deleteFile(dir, oldPaint.imageFilename);
         fileInfo = await resizeAndSaveImage(newFile, title, dir);
       }
 
@@ -82,7 +82,7 @@ export async function updatePainting(
                 id: Number(rawFormData.categoryId),
               },
             }
-          : oldPaint.categoryId !== null
+          : oldPaint.categoryId
             ? {
                 disconnect: {
                   id: oldPaint.categoryId,
@@ -122,7 +122,7 @@ export async function deletePainting(id: number) {
       where: { id },
     });
     if (painting) {
-      deleteFile(dir, painting.images[0].filename);
+      deleteFile(dir, painting.imageFilename);
       await prisma.painting.delete({
         where: {
           id,
@@ -138,9 +138,22 @@ export async function deletePainting(id: number) {
 
 export async function deleteCategoryPainting(id: number) {
   try {
-    await prisma.paintingCategory.delete({
+    const cat = await prisma.paintingCategory.findUnique({
       where: { id },
     });
+
+    if (cat) {
+      const contentId = cat.categoryContentId;
+      if (contentId) {
+        await prisma.categoryContent.delete({
+          where: { id: contentId },
+        });
+      }
+
+      await prisma.paintingCategory.delete({
+        where: { id },
+      });
+    }
     revalidatePath("/admin/peintures");
     return { message: "Catégorie supprimée", isError: false };
   } catch (e) {
@@ -152,20 +165,30 @@ export async function createCategoryPainting(
   prevState: { message: string; isError: boolean } | null,
   formData: FormData,
 ) {
-  try {
-    const value = formData.get("text") as string;
-    const key = transformValueToKey(value);
+  const rawFormData = Object.fromEntries(formData);
+  const value = rawFormData.value as string;
 
+  try {
     await prisma.paintingCategory.create({
       data: {
-        key,
+        key: transformValueToKey(value),
         value,
+        content: {
+          create: {
+            title: rawFormData.title as string,
+            text: rawFormData.text as string,
+            imageFilename: rawFormData.filename as string,
+            imageWidth: Number(rawFormData.width),
+            imageHeight: Number(rawFormData.height),
+          },
+        },
       },
     });
+
     revalidatePath("/admin/peintures");
     return { message: "Catégorie ajoutée", isError: false };
   } catch (e) {
-    return { message: "Erreur à la création", isError: true };
+    return { message: `Erreur à la création`, isError: true };
   }
 }
 
@@ -173,22 +196,48 @@ export async function updateCategoryPainting(
   prevState: { message: string; isError: boolean } | null,
   formData: FormData,
 ) {
-  try {
-    const rawFormData = Object.fromEntries(formData);
-    const id = Number(rawFormData.id);
-    const value = rawFormData.text as string;
-    const key = transformValueToKey(value);
+  const rawFormData = Object.fromEntries(formData);
+  const id = Number(rawFormData.id);
+  const value = rawFormData.value as string;
 
-    await prisma.paintingCategory.update({
+  try {
+    const oldCat = await prisma.paintingCategory.findUnique({
       where: { id },
-      data: {
-        key,
-        value,
-      },
+      include: { content: true },
     });
+
+    if (oldCat) {
+      let content;
+      const data = {
+        title: rawFormData.title as string,
+        text: rawFormData.text as string,
+        imageFilename: rawFormData.filename as string,
+        imageWidth: Number(rawFormData.width),
+        imageHeight: Number(rawFormData.height),
+      };
+
+      if (!oldCat.content) {
+        content = {
+          create: data,
+        };
+      } else {
+        content = {
+          update: data,
+        };
+      }
+
+      await prisma.paintingCategory.update({
+        where: { id },
+        data: {
+          key: transformValueToKey(value),
+          value,
+          content,
+        },
+      });
+    }
     revalidatePath("/admin/peintures");
     return { message: "Catégorie modifiée", isError: false };
   } catch (e) {
-    return { message: "Erreur à la modification", isError: true };
+    return { message: `Erreur à la modification`, isError: true };
   }
 }
